Remove need for SiegeWarItemUseController.
