+++ src/com/palmergames/bukkit/towny/war/siegewar/listeners/SiegeWarActionListener.java
@@ -3,21 +3,28 @@ package com.palmergames.bukkit.towny.war.siegewar.listeners;
 import java.util.ArrayList;
 import java.util.List;
 
+import org.bukkit.Material;
 import org.bukkit.block.Block;
 import org.bukkit.event.EventHandler;
+import org.bukkit.event.EventPriority;
 import org.bukkit.event.Listener;
+import org.bukkit.event.player.PlayerItemConsumeEvent;
 
 import com.palmergames.bukkit.towny.Towny;
+import com.palmergames.bukkit.towny.TownyMessaging;
 import com.palmergames.bukkit.towny.TownySettings;
+import com.palmergames.bukkit.towny.TownyUniverse;
+import com.palmergames.bukkit.towny.event.actions.TownyBuildEvent;
 import com.palmergames.bukkit.towny.event.actions.TownyBurnEvent;
 import com.palmergames.bukkit.towny.event.actions.TownyDestroyEvent;
 import com.palmergames.bukkit.towny.event.actions.TownyExplodingBlocksEvent;
 import com.palmergames.bukkit.towny.object.Translation;
+import com.palmergames.bukkit.towny.war.siegewar.objects.Siege;
 import com.palmergames.bukkit.towny.war.siegewar.utils.SiegeWarBlockUtil;
+import com.palmergames.bukkit.towny.war.siegewar.utils.SiegeWarDistanceUtil;
 
 public class SiegeWarActionListener implements Listener {
 
-	@SuppressWarnings("unused")
 	private final Towny plugin;
 	
 	public SiegeWarActionListener(Towny instance) {
@@ -51,6 +58,42 @@ public class SiegeWarActionListener implements Listener {
 				}
 			}
 			event.setBlockList(filteredList);
+		}
+	}
+	
+	@EventHandler(priority = EventPriority.NORMAL, ignoreCancelled = true)
+	public void onPlayerConsume(PlayerItemConsumeEvent event) {
+		if (plugin.isError()) {
+			event.setCancelled(true);
+			return;
+		}
+
+		if(TownySettings.getWarSiegeEnabled()) {
+			try {
+				//Prevent milk bucket usage while attempting to gain banner control
+				if(event.getItem().getType() == Material.MILK_BUCKET) {
+					for(Siege siege: TownyUniverse.getInstance().getDataSource().getSieges()) {
+						if(siege.getBannerControlSessions().containsKey(event.getPlayer())) {
+							event.setCancelled(true);
+							TownyMessaging.sendErrorMsg(event.getPlayer(), Translation.of("msg_war_siege_zone_milk_bucket_forbidden_while_attempting_banner_control"));
+						}
+					}
+				}
+		
+			} catch (Exception e) {
+				System.out.println("Problem evaluating siege player consume event");
+				e.printStackTrace();
+			}
+		}
+	}
+	
+	@EventHandler
+	public void onBucketUse(TownyBuildEvent event) {
+		if(TownySettings.getWarSiegeEnabled() && event.isInWilderness() && TownySettings.isWarSiegeZoneBucketEmptyingRestrictionsEnabled() && TownySettings.getWarSiegeZoneBucketEmptyingRestrictionsMaterials().contains(event.getMaterial())) {
+			if (SiegeWarDistanceUtil.isLocationInActiveSiegeZone(event.getLocation())) {
+				event.setMessage(Translation.of("msg_war_siege_zone_bucket_emptying_forbidden"));
+				event.setCancelled(true);
+			}
 		}
 	}
 }
